refactor: move top header to separate file
