+++ pages/index.tsx
@@ -13,33 +13,20 @@ import Button from '@mui/material/Button'
 import Dialog from '@mui/material/Dialog'
 
 // Marketplace components
-import SignUp from '../components/SignUp'
-import SignIn from '../components/SignIn'
-import ResetPassword from '../components/ResetPassword'
-import UserDetails from '../components/UserDetails'
+import TopMenu from '../components/TopMenu'
 
-// Variables
-const color = 'white'
+// nhost
+import { gql } from '@apollo/client'
 
-// Event handlers
-function handleSearch(event: React.FormEvent<HTMLFormElement>) {
-    event.preventDefault()
-    console.log('searching, event: ', event)
-}
+const GET_LISTINGS = gql`
+  query GetListings() {
+    listings{
+      *
+    }
+  }
+`
 
 export default function Home() {
-    const [showLoginModal, setShowLoginModal] = React.useState(false)
-    const [loginModalTab, setLoginModalTab] = React.useState<
-        'signup' | 'signin' | 'resetpassword'
-    >('signup')
-
-    function handleMenuLoginButtonClick(
-        event: React.MouseEvent<HTMLButtonElement>,
-    ) {
-        // console.log('Login button clicked, event: ', event)
-        setShowLoginModal(true)
-    }
-
     return (
         <>
             <Head>
@@ -54,48 +41,12 @@ export default function Home() {
                 />
                 <link rel="icon" href="/favicon.ico" />
             </Head>
-            <div
-                css={{
-                    padding: '32px',
-                    'background-color': 'turquoise',
-                    '&:hover': {
-                        color: color,
-                    },
-                }}
-            >
-                <h1>Marketplace</h1>
-                (login, see favourite listings)
-                <Button variant="text" onClick={handleMenuLoginButtonClick}>
-                    Login
-                </Button>
-            </div>
-            <main>
-                <form onSubmit={handleSearch}>
-                    <input type="text" placeholder="Search" />
-                    <button>Search</button>
-                </form>
 
+            <TopMenu />
+
+            <main>
                 <p>Recent/relevant for you offers go here</p>
             </main>
-            <Dialog
-                onClose={() => setShowLoginModal(false)}
-                open={showLoginModal}
-            >
-                <div>Modal goes here</div>
-                <Button onClick={() => setLoginModalTab('signup')}>
-                    Sign up
-                </Button>
-                <Button onClick={() => setLoginModalTab('signin')}>
-                    Sign in
-                </Button>
-                <Button onClick={() => setLoginModalTab('resetpassword')}>
-                    Reset password
-                </Button>
-                {loginModalTab === 'signup' && <SignUp />}
-                {loginModalTab === 'signin' && <SignIn />}
-                {loginModalTab === 'resetpassword' && <ResetPassword />}
-                <UserDetails />
-            </Dialog>
         </>
     )
 }
